requirements.txt
# 【金融汇报】(分析主理人) 核心依赖
ccxt>=4.0.0
agno>=0.1.0
python-dotenv>=1.0.0

# 【核心脚本】(底层工具箱) 依赖
watchdog>=3.0.0

# 通用工具
requests>=2.31.0
